Separated code into lib and main for integration tests
