+++ src/lib/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod system;
+pub mod types;
